backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.0
python-multipart>=0.0.7
openpyxl>=3.1.0
pandas>=2.2.0
celery>=5.3.0
redis>=5.0.0
